Update README and requirements.txt
Ran current code through pylint to format everything correctly. In its current state, the federation_service does perform its role in the most basic way of aggregating all the responses back from a given service. No higher level processing is applied to the federated queries at this time, differing the responseObject generated here compared to CanDIGv1. The merge_count function is not yet fully implemented due to differences in the way data is represented and needs to either be addressed or written off.